+++ src/main/java/me/zeroeightsix/kami/module/modules/misc/AutoFish.kt
@@ -5,8 +5,8 @@ import me.zeroeightsix.kami.event.events.SafeTickEvent
 import me.zeroeightsix.kami.mixin.extension.rightClickMouse
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.setting.Settings
-import me.zeroeightsix.kami.util.BlockUtils.isWater
 import me.zeroeightsix.kami.util.TickTimer
+import me.zeroeightsix.kami.util.WorldUtils.isWater
 import net.minecraft.init.Items
 import net.minecraft.network.play.server.SPacketSoundEffect
 import org.kamiblue.event.listener.listener
